Fix #319 Exception when parent invalid
Existing implementation that used `.Substring($"{baseUrl}/{customerId}")` was prone to errors and potentially unpredictable behaviour in some potential parent/public id's. It did however prevent in most cases a misidentification of a path with one of a different customer.

I have expanded on this strength by checking separately for a matching host and separately performing "check and trim" of the path in the form of `/{customerId}/a/path/to/something/`.